% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateVALUE.R
\name{aggregateVALUE}
\alias{aggregateVALUE}
\title{Aggregation of value objects along selected dimensions}
\usage{
aggregateVALUE(valueObj, aggr.m = list(FUN = "mean", na.rm = TRUE))
}
\arguments{
\item{valueObj}{A VALUE object to be aggregated.}

\item{aggr.m}{Monthly aggregation function. A list indicating the name of the
aggregation function in first place, and other optional arguments to be passed to the aggregation function. See the examples.}
}
\value{
A VALUE object aggregated along the chosen dimension(s).
}
\description{
Aggregates a grid along the target dimensions through aggregation function specification.
}
\details{
Currently only the temporal monthly aggregation is implemented
}
\examples{

obs.file <- file.path(find.package("R.VALUE"),
                      "example_datasets",
                      "VALUE_ECA_86_v2.zip")
tmin.DJF.daily <- loadValueStations(obs.file, "tmin",
                                    season = c(12,1,2), years = 1970:2008)
# Monthly means:
tmin.DJF.monthlyMean <- aggregateVALUE(tmin.DJF.daily, aggr.m = list(FUN = "mean", na.rm = TRUE))
# Other functions can be introduced by the user. E.g., monthly median:
tmin.DJF.monthlyMedian <- aggregateVALUE(tmin.DJF.daily,
                                       aggr.m = list(FUN = "quantile", probs = .5, na.rm = TRUE))
}
\author{
J. Bedia
}

