% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrMat.VALUE.R
\name{corrMat.VALUE}
\alias{corrMat.VALUE}
\title{Correlation matrix for paper Figures 4 and 5 on spatial validation}
\usage{
corrMat.VALUE(stationObj, predictions.file = NULL, season = c("annual",
  "DJF", "MAM", "JJA", "SON"), method = c("pearson", "kendall", "spearman"),
  type = c("before", "after"), max.na.prop = 0.25, deseason = TRUE,
  window.width = 60)
}
\arguments{
\item{stationObj}{An R-VALUE object containing station data (as returned by \code{\link{loadValueStations}}).}

\item{predictions.file}{Path to the file storing the predictions (passed to \code{\link{loadValuePredictions}}).
Default to NULL, meaning that the correlation matrix is done on the observations.}

\item{season}{Character string indicating the target season. Accepted values are 
\code{c("annual", "DJF", "MAM", "JJA", "SON")}}

\item{method}{Correlation method, passed to the \code{\link{cor}} function}

\item{type}{Type of aggregation in the case of multiple realizations. Should the aggregation of 
multiple members be performed \code{"before"} or \code{"after"} computing the correlations?. Ignored
 in the case of observations and deterministic predictions.}

\item{max.na.prop}{Maximum allowed proportion of missing data (Default to 0.25). See details}

\item{deseason}{Logical. Should the seasonal cycle from the input data series be removed?. See details.}

\item{window.width}{Used if \code{deseason = TRUE}. Integer number indicating the width, in days, of the window used for
moving average computation of the reference daily climatology. Default to 60 days. See details.}
}
\value{
A 2D matrix. Attributes indicate the station names (in the row/column order they appear), and their
geographical coordinates.
}
\description{
Computes the cross correlation matrices between stations that serve as input for plotting functions
}
\details{
A (circular) moving average daily climatology can be automatically calculated for each data series, considering 
a specific window width (60 days by default, centered around lag 0). This is used for removing the seasonal cycles when applicable. 
An additional argument allows to specify the maximum number of missing values allowed, being ommited from the analysis 
those data series above the threshold.
}
\author{
J. Bedia
}

