% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseason.VALUE.R
\name{deseason.VALUE}
\alias{deseason.VALUE}
\title{Remove seasonal cycle}
\usage{
deseason.VALUE(valueObj, window.width = 31, max.na.prop = 0.25)
}
\arguments{
\item{valueObj}{Any VALUE object (as returned either by \code{\link{loadValueStations}} or \code{\link{loadValuePredictions}}).}

\item{window.width}{Used if \code{deseason = TRUE}. Integer number indicating the width, in days, of the window used for
moving average computation of the reference daily climatology. Default to 31 days. See details.}

\item{max.na.prop}{Maximum allowed proportion of missing data (Default to 0.25). See details}
}
\value{
A VALUE object same as input but with seasonal cycles removed.
}
\description{
Remove seasonal cycle from a VALUE dataset using a moving average filter
}
\details{
A (circular) moving average daily climatology can be automatically calculated for each data series, considering 
a user-defined window width centered around lag 0. An additional argument allows to specify the maximum number of missing values allowed,
 being ommited from the analysis those data series above the threshold.

A global attribute "deseason::window.width" is appended to the output, indicating the window width (in days).
}
\note{
Although the function is insentitive to the input season, it is recommended to perform seasonal cycle removal on
the annual series, rather than in particular seasons. Seasonal slices can be selected afterwards using \code{\link{subsetVALUE}}.
}
\examples{
\dontrun{
obs.file <- file.path(find.package("R.VALUE"),"example_datasets","VALUE_ECA_86_v2.zip")
# Load 1 station (Innsbruck) as example:
stationObj <- loadValueStations(obs.file, var = "tmin", stationID = "000013", years = 1991:1993)
# Test different window sizes:
widths <- c(31,45,60)
out <- lapply(widths, function(x) deseason.VALUE(stationObj, window.width = x))
dates <- as.Date(stationObj$Dates$start)
plot(dates, stationObj$Data, ty = 'l', ylab = "tmin (degC)")
for (i in 1:length(widths)) {
      lines(dates,out[[i]]$Data,col = i + 1)
}
legend("bottomright", legend = widths, lty = 1, col = (1:3)+1, title = "window.width")
}
}
\author{
J. Bedia
}

