% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFUN.R
\name{dimFix}
\alias{dimFix}
\title{Complete missing dimensions of VALUE objects}
\usage{
dimFix(valueObj)
}
\arguments{
\item{A}{VALUE R object}
}
\value{
The same object with all the dimensions (i.e. member, time, station)
}
\description{
Inverse of drop to complete all dimensions of the Data array
}
\author{
J. Bedia
}
\keyword{internal}

