% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getACF}
\alias{getACF}
\title{Autocorrelation}
\usage{
getACF(data, lag.max)
}
\description{
Get annual and seasonal Autocorrelation at pre-defined lags from a station or field object
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, S. Herrera and J. Bedia
}
\keyword{internal}

