% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{getAmountFreqGT}
\alias{getAmountFreqGT}
\title{Amount from events above threshold}
\usage{
getAmountFreqGT(data, threshold, MARGIN = NULL)
}
\description{
Get annual and seasonal amount falling in heavy rainy days from a station or field object
}
\details{
Especific function for precipitation
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg} and J. Bedia and S. Herrera
}
\keyword{internal}

