% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAnnualCicle}
\alias{getAnnualCicle}
\title{Annual Cycle}
\usage{
getAnnualCicle(data, INDEX = 1:dim(data)[1])
}
\description{
Function to compute VALUE indices for mean annual cycle
}
\author{
Sven Kotlarski \email{sven.kotlarski@env.ethz.ch}, J. Bedia and S. Herrera
}
\keyword{internal}

