% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{getCorrelation}
\alias{getCorrelation}
\title{Correlation}
\usage{
getCorrelation(o, p, MARGIN, method)
}
\description{
Get correlation
}
\author{
Daniel San-Mart\'in \email{daniel@predictia.es}
}
\keyword{internal}

