% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFreqDW}
\alias{getFreqDW}
\title{Dry-wet probability}
\usage{
getFreqDW(data, threshold)
}
\description{
Get annual and seasonal dry-wet-probability from a station or field object
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg} and J. Bedia and S. Herrera
}
\keyword{internal}

