% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFreqGET}
\alias{getFreqGET}
\title{Get annual and seasonal amount of days greater or equal than a predefined threshold from a station or field object}
\usage{
getFreqGET(data, threshold, MARGIN = NULL)
}
\description{
Function to calculate the average number of days/season  with measurable precipitation (e.g. precipitation > 1mm).
}
\author{
Elke Hertig \email{elke.hertig@geo.uni-augsburg.de} and J. Bedia
}
\keyword{internal}

