% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGTsld}
\alias{getGTsld}
\title{Spell length distribution}
\usage{
getGTsld(data, threshold, INDEX = 1:dim(data)[1])
}
\description{
This function calculates pre-defined quantiles of a spell-distribution and the median of annual maximum consecutive spells
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg} and J. Bedia and S. Herrera
}
\keyword{internal}

