% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getIntersect}
\alias{getIntersect}
\title{getIntersect}
\usage{
getIntersect(obs, prd)
}
\arguments{
\item{obs}{Value object of observations}

\item{prd}{Value object of predictions}
}
\value{
A list with obj and pred intersected
}
\description{
Temporal and spatial matching between obs and pred
}
\author{
S. Herrera
}
\keyword{internal}

