% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{getLTsld}
\alias{getLTsld}
\title{Get annual and seasonal Below of a predefined threshold Spell Length Distribution from a station or field object}
\usage{
getLTsld(data, threshold, INDEX = 1:dim(data)[1])
}
\description{
Get annual and seasonal Below of a predefined threshold Spell Length Distribution from a station or field object
}
\keyword{internal}

