% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{getMean}
\alias{getMean}
\title{Mean}
\usage{
getMean(data, MARGIN)
}
\description{
Get annual and seasonal mean from a station or field object
}
\author{
Ole Roessler \email{ole.roessler@giub.unibe.ch} and J. Bedia
}
\keyword{internal}

