% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getReturnValue}
\alias{getReturnValue}
\title{Return Value}
\usage{
getReturnValue(data, prob, INDEX = 1:dim(data)[1])
}
\description{
Get annual and seasonal 1/prob return value (left/right tail) from a station or field object. Require package - evd
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg} and J. Bedia
}
\keyword{internal}

