% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSkew}
\alias{getSkew}
\title{Get annual and seasonal skewness from a station or field object}
\usage{
getSkew(data, MARGIN)
}
\description{
Get annual and seasonal skewness from a station or field object
}
\author{
Ole Roessler \email{ole.roessler@giub.unibe.ch} and J. Bedia
}
\keyword{internal}

