% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadValueStations.R
\name{getTimeDomainValueStations}
\alias{getTimeDomainValueStations}
\title{Time  index positions for station dataset selections}
\usage{
getTimeDomainValueStations(timeDates, season, years)
}
\arguments{
\item{timeDates}{a POSIXlt vector of time dates}

\item{season}{A vector of months defining the season selected}

\item{years}{A vector of (continuous) year selection}
}
\value{
A list with a vector of time index positions and the corresponding POSIXlt dates
}
\description{
Get time index positions for loading VALUE ascii station data
}
\author{
J. Bedia
}
\keyword{internal}
