% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.acf.R
\name{index.acf}
\alias{index.acf}
\title{Autocorrelation function}
\usage{
index.acf(ts, lag.max = 3)
}
\arguments{
\item{ts}{A vector containing the time series}

\item{lag.max}{Maximum lag considered for acf calculation. Default \code{lag.max = 3}.}
}
\value{
A vector with the estimated autocorrelation values.
}
\description{
Function to compute the estimated autocorrelation function at pre-defined lags.
}
\examples{
\dontrun{
# Autocorrelation of Braganca:
data(precipIberiaECA)
obs <- precipIberiaECA$observations$Data[,1]
plot(index.acf(obs, lag.max = 10), type = "h")
}
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, J. Bedia, D. San-Mart\'in, S. Herrera
}

