% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.amountFreqGE.R
\name{index.amountFreqGE}
\alias{index.amountFreqGE}
\title{Amount from events above threshold}
\usage{
index.amountFreqGE(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the input time series}

\item{threshold}{A float number defining the threshold considered. Default to 1.}
}
\value{
A float number corresponding to the amount of days above the defined threshold of the input.
}
\description{
Function to compute the amount from events above or equal a predefined threshold.
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, J. Bedia, D. San-Mart\'in, S. Herrera
}

