% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.annual.cycle.filtered.R
\name{index.annual.cycle.filtered}
\alias{index.annual.cycle.filtered}
\title{Filtered annual cycle}
\usage{
index.annual.cycle.filtered(ts, dates, peak = 1, type = "max",
  thresh = 0.05)
}
\arguments{
\item{ts}{A vector containing the data}

\item{dates}{A character (or \code{POSIXct}) vector following the format \dQuote{YYYY-MM-DD}
(i.e., \code{format = "\%Y-\%m-\%d"} as in \code{\link{strptime}}). This is directly passed by the
 VALUE objects (either stations or predictions) through the element \code{object$Dates$start}.}

\item{peak}{= 1 or 2}

\item{type}{A character string indicating max (default), min phase (days), amp (amplitude) or amprel (relative amplitude)}

\item{thresh}{Threshold that defines weak maxima (relative to amplitude of annual cycle)}
}
\value{
A float number corresponding to the amplitude or the (phase) day of the maxima
}
\description{
Returns the amplitude or the location of the maxima of the annual cycle
}
\author{
Douglas Maraun, J. Bedia, D. San-Martin, JM. Gutierrez
}
