% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freq.R
\name{index.freq}
\alias{index.freq}
\title{Relative/Absolute frequencies w.r.t. user-defined thresholds}
\usage{
index.freq(ts, threshold = 1, condition, freq.type)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered. Default to 1 (for precipitation).}

\item{condition}{Inequality operator to be applied considering the given threshold.
\code{"GT"} = greater than the value of \code{threshold}, \code{"GE"} = greater or equal,
  \code{"LT"} = lower than, \code{"LE"} = lower or equal than.}

\item{freq.type}{A character string indicating if frequecies to be computed are relative (\code{"rel"}) or
absolute (counts) (\code{"abs"}).}
}
\value{
A float number corresponding to the proportion/number of days below/above the defined threshold.
}
\description{
Function to compute the relative/absolute amount of days below/above a predefined threshold.
}
\author{
Douglas Maraun, J. Bedia, D. San-Martin, S. Herrera
}
