% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freq.rel.GE.R
\name{index.freq.rel.GE}
\alias{index.freq.rel.GE}
\title{Realtive fequency above or equal the threshold}
\usage{
index.freq.rel.GE(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered. Default to 1.}
}
\value{
A float number corresponding to the proportion of days above or equal the defined threshold of the input.
}
\description{
Function to compute the proportion of days above or equal a predefined threshold mean index. Missing values are ignored.
}
\author{
Douglas Maraun, J. Bedia, D. San-Martin, S. Herrera
}

