% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freq.rel.GT.R
\name{index.freq.rel.GT}
\alias{index.freq.rel.GT}
\title{Relative frequency above the threshold}
\usage{
index.freq.rel.GT(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered. Default to 1.}
}
\value{
A float number corresponding to the proportion of days above the defined threshold of the input.
}
\description{
Function to compute the proportion of days above a predefined threshold.
}
\author{
Douglas Maraun, J. Bedia, D. San-Martin, S. Herrera
}

