% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freq.rel.LE.R
\name{index.freq.rel.LE}
\alias{index.freq.rel.LE}
\title{Relative frequency lower or equal the threshold}
\usage{
index.freq.rel.LE(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered. Default to 1}
}
\value{
A float number corresponding to the proportion of days below or equal the defined threshold of the input.
}
\description{
Function to compute the proportion of days below or equal of a predefined threshold mean index.
}
\author{
Douglas Maraun
}

