% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freq.rel.LT.R
\name{index.freq.rel.LT}
\alias{index.freq.rel.LT}
\title{Relative frequency below the threshold}
\usage{
index.freq.rel.LT(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered. Default to 1.}
}
\value{
A float number corresponding to the proportion of days below the defined threshold of the input.
}
\description{
Function to compute the relative amount of days below a predefined threshold.
}
\author{
Douglas Maraun, J. Bedia, D. San-Martin, S. Herrera
}

