% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freqDD.R
\name{index.freqDD}
\alias{index.freqDD}
\title{Dry-dry probability}
\usage{
index.freqDD(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered. Default to 1.}
}
\value{
A float number corresponding to the dry-dry transition probability.
}
\description{
Function to compute the dry-dry probability index.
}
\examples{
\dontrun{
# Dry-dry transition probability of Braganca:
data(precipIberiaECA)
obs <- precipIberiaECA$observations$Data[,1]
# Default precipitation threshold of 1 mm
index.freqDD(obs)
}
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, J. Bedia, D. San-Mart\'in, S. Herrera
}

