% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freqDW.R
\name{index.freqDW}
\alias{index.freqDW}
\title{Dry-wet probability}
\usage{
index.freqDW(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered. Default to 1.}
}
\value{
A float number corresponding to the dry-wet transition probability.
}
\description{
Function to compute the dry-wet probability index. Missing values are ignored.
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, J. Bedia, D. San-Mart\'in, S. Herrera
}

