% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freqGT.R
\name{index.freqGT}
\alias{index.freqGT}
\title{Frequency above the threshold}
\usage{
index.freqGT(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered. Default to 1.}
}
\value{
A float number corresponding to the amount of days above the defined threshold of the input.
}
\description{
Function to compute the amount of days above a predefined threshold.
}
\author{
Dougals Maraun \email{dmaraun@geomar.de}
}

