% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freqLT.R
\name{index.freqLT}
\alias{index.freqLT}
\title{Frequency below the threshold}
\usage{
index.freqLT(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered. Default to 1.}
}
\value{
A float number corresponding to the amount of days below the defined threshold of the input.
}
\description{
Function to compute the amount of days below a predefined threshold.
}
\author{
Douglas Maraun \email{dmaraun@geomar.de}, J. Bedia, D. San-Mart\'in, S. Herrera
}

