% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freqWD.R
\name{index.freqWD}
\alias{index.freqWD}
\title{Wet-dry probability}
\usage{
index.freqWD(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered.}
}
\value{
A float number corresponding to the wet-dry transition probability.
}
\description{
Function to compute the wet-dry probability index.
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, J. Bedia, D. San-Mart\'in, S. Herrera
}

