% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.lfvar.R
\name{index.lfvar}
\alias{index.lfvar}
\title{Proportion of variance in low frequency}
\usage{
index.lfvar(ts, lowVarPeriod, INDEX = 1:length(ts))
}
\arguments{
\item{ts}{A vector containing the data}

\item{lowVarPeriod}{Low frequency threshold. Minimum cycles per unit time below which frequency is considered as low}

\item{INDEX}{A Vector defining the aggregation periods (passed to \code{tapply}). No aggregation is performed by default.}
}
\description{
Annual and seasonal proportion of variance in low frequency
}
\author{
Douglas Maraun \email{dmaraun@geomar.de}
}

