% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.mean.R
\name{index.mean}
\alias{index.mean}
\title{Arithmetic mean}
\usage{
index.mean(ts)
}
\arguments{
\item{ts}{A vector containing the data}
}
\value{
A float number corresponding to the mean of the input.
}
\description{
Function to compute the arithmetic mean index.
}
\examples{
\dontrun{
# Precipitation mean of Braganca:
data(precipIberiaECA)
obs <- precipIberiaECA$observations$Data[,1]
index.mean(obs)
}
}
\author{
Ole Roessler \email{ole.roessler@giub.unibe.ch}, J. Bedia, D. San-Mart\'in, S. Herrera
}

