% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.meanGE.R
\name{index.meanGE}
\alias{index.meanGE}
\title{Arithmetic mean of days equal or above threshold}
\usage{
index.meanGE(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{Threshold value. Default to 1.}
}
\value{
A float number corresponding to the mean of days above the threshold.
}
\description{
Function to compute the arithmetic mean of days equal or above a given threshold.
}
\details{
The function is envisaged for calculating the mean precipitation of wet days, excluding dry ones
 (\emph{a.k.a.} simple day intensity index).
}
\author{
J. Bedia
}

