% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.prcXXth.R
\name{index.prcXXth}
\alias{index.prcXXth}
\title{Sample quantiles}
\usage{
index.prcXXth(ts, prob = seq(0, 1, 0.25), threshold = NULL)
}
\arguments{
\item{ts}{A vector containing the data}

\item{prob}{A float number in the range [0,1] defining the probability of the quantile to be calculated.}

\item{threshold}{Optional parameter. Remove data below the threshold value.
Included to consider the case of precipitation. NULL by default.}
}
\value{
A float number corresponding to the quantile.
}
\description{
Function to compute the sample quantiles for the given probabilities.
}
\examples{
\dontrun{
# 98th quantile of Braganca:
data(precipIberiaECA)
obs <- precipIberiaECA$observations$Data[,1]
# 98th percentile
prob <- 0.98
# Default 
index.prcXXth(obs)
}
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, J. Bedia, D. San-Mart\'in, S. Herrera
}

