% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.rv.lb.R
\name{index.rv.lb}
\alias{index.rv.lb}
\title{Return Value lower tail}
\usage{
index.rv.lb(ts, prob = 20, INDEX = 1:length(ts))
}
\arguments{
\item{ts}{A vector containing the data}

\item{prob}{Return period considered. Default to 20.}

\item{INDEX}{A Vector defining the aggregation periods (passed to \code{tapply}). No aggregation is performed by default.}
}
\value{
Return value for the lower tail
}
\description{
Function to compute the 1/prob return value (left tail).
}
\examples{
\dontrun{
# 20-years Return value of Braganca:
data(precipIberiaECA)
obs <- precipIberiaECA$observations$Data[,1]
index.rv.lb(obs, prob = 20)
}
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, J. Bedia, D. San-Mart\'in, S. Herrera
}

