% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.rv.R
\name{index.rv}
\alias{index.rv}
\title{Return Value}
\usage{
index.rv(ts, prob, INDEX = 1:length(ts))
}
\arguments{
\item{ts}{A vector containing the data}

\item{prob}{Return period considered}

\item{INDEX}{A Vector defining the aggregation periods (passed to \code{tapply}). No aggregation is performed by default.}
}
\value{
Return values for the upper/lower tails
}
\description{
Function to compute the 1/prob return value (left/right tail).
}
\examples{
\dontrun{
# 20-years Return value of Braganca:
data(precipIberiaECA)
obs <- precipIberiaECA$observations$Data[,1]
index.rv(obs, prob = 20)
}
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, J. Bedia, D. San-Mart\'in, S. Herrera
}

