% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.sd.R
\name{index.sd}
\alias{index.sd}
\title{Standard deviation}
\usage{
index.sd(ts)
}
\arguments{
\item{ts}{A vector containing the data}
}
\value{
A float number corresponding to the standard deviation of the input sample.
}
\description{
Function to compute the standard deviation index.
}
\examples{
\dontrun{
# Precipitation standard deviation of Braganca:
data(precipIberiaECA)
obs <- precipIberiaECA$observations$Data[,1]
index.sd(obs)
}
}
\author{
Ole Roessler \email{ole.roessler@giub.unibe.ch}, J. Bedia, D. San-Mart\'in, S. Herrera
}

