% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.skew.R
\name{index.skew}
\alias{index.skew}
\title{Skewness}
\usage{
index.skew(ts)
}
\arguments{
\item{ts}{A vector containing the data}
}
\value{
A float number corresponding to the skewness index.
}
\description{
Function to compute the skewness index.
}
\author{
Ole Roessler \email{ole.roessler@giub.unibe.ch}, J. Bedia, D. San-Mart\'in, S. Herrera
}

