% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.spellmax.annual.prcXXth.R
\name{index.spellmax.annual.prcXXth}
\alias{index.spellmax.annual.prcXXth}
\title{Annual spell length percentiles}
\usage{
index.spellmax.annual.prcXXth(ts, dates, threshold = 1,
  threshold.type = "abs", condition = c("GT", "GE", "LT", "LE"),
  prob = 0.5)
}
\arguments{
\item{ts}{A vector containing the data}

\item{dates}{A character (or \code{POSIXct}) vector following the format \dQuote{YYYY-MM-DD}
(i.e., \code{format = "\%Y-\%m-\%d"} as in \code{\link{strptime}}). This is directly passed by the
 VALUE objects (either stations or predictions) through the element \code{object$Dates$start}.}

\item{threshold}{A float number defining the threshold considered. Default to 1}

\item{threshold.type}{Is the value of \code{threshold} an absolute value or a quantile [0-1]?.
Two possible values: \code{"abs"} and \code{"prob"} respectively. Default to \code{"abs"}.}

\item{condition}{Inequality operator to be applied considering the given threshold.
\code{"GT"} = greater than the value of \code{threshold}, \code{"GE"} = greater or equal,
 \code{"LT"} = lower than, \code{"LE"} = lower or equal than}

\item{prob}{A float number in the range [0,1] defining the probability of the quantile to be calculated.
Default to median (\code{prob = 0.5}).}
}
\value{
A float number with the corresponding percentile.
}
\description{
Computes a user-defined percentile for the duration of spells (above/below) a given threshold, on an annual basis.
}
\details{
The function requires the date information in order to ensure that spells
 are computed on consecutive records.
}
\author{
J. Bedia, J.M. Gutierrez
}
