% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.sum.R
\name{index.sum}
\alias{index.sum}
\title{Sum}
\usage{
index.sum(ts)
}
\arguments{
\item{ts}{A vector containing the data}
}
\value{
A float number corresponding to the sum of the input.
}
\description{
Function to compute the sum index.
}
\author{
Ole Roessler \email{ole.roessler@giub.unibe.ch}, J. Bedia, D. San-Mart\'in, S. Herrera
}

