% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.trend.R
\name{index.trend}
\alias{index.trend}
\title{Interannual trend}
\usage{
index.trend(ts, dates, output = c("coef", "sig"), sig.level = 0.95)
}
\arguments{
\item{ts}{A vector containing the data}

\item{dates}{A character (or \code{POSIXct}) vector following the format \dQuote{YYYY-MM-DD}
(i.e., \code{format = "\%Y-\%m-\%d"} as in \code{\link{strptime}}). This is directly passed by the
 VALUE objects (either stations or predictions) through the element \code{object$Dates$start}.}

\item{output}{Desired output. The options are \code{"coef"} for the slope
coefficient (default) or \code{"sig"}, which is a binary value indicating whether 
the slope coefficient is significant (1) or not (0) at the given \code{sig.level}.}

\item{sig.level}{Significance level. Default to 0.95.}
}
\value{
A numeric value with either the slope coefficient (for \code{output = "coef"})
or whether it is significant or not (\code{output = "sig"}) at the specified \code{sig.level}
}
\description{
Computes slope and significance of the interannual linear trend
}
\author{
J Bedia
}
