% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.var.R
\name{index.var}
\alias{index.var}
\title{Variance}
\usage{
index.var(ts)
}
\arguments{
\item{ts}{A vector containing the data}
}
\value{
A float number corresponding to the variance of the input.
}
\description{
Function to compute the variance index.
}
\author{
Ole Roessler \email{ole.roessler@giub.unibe.ch}, J. Bedia, D. San-Mart\'in, S. Herrera
}
