% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.wetAmount.prcXXth.R
\name{index.wetAmount.prcXXth}
\alias{index.wetAmount.prcXXth}
\title{Total amount of precipitation of wet days above percentile}
\usage{
index.wetAmount.prcXXth(ts, prob = 0.98)
}
\arguments{
\item{ts}{A vector containing the data}

\item{prob}{A float number in the range [0,1] defining the probability of the quantile to be calculated.
Default to 0.98}
}
\value{
A float number corresponding to the proportion/number of days below/above the defined threshold.
}
\description{
Function to compute the total amount of precipitation falling in wet days above percentile.
}
\details{
Wet days are internally defined as those exceeding the 1mm threshold
}
\author{
J. Bedia
}

