% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointProbMat.VALUE.R
\name{jointProbMat.VALUE}
\alias{jointProbMat.VALUE}
\title{Joint probability analysis}
\usage{
jointProbMat.VALUE(stationObj, predictions.file = NULL, season = c("annual",
  "DJF", "MAM", "JJA", "SON"), aggr.type = c("before", "after"),
  prob.type = c("DD", "DW", "WW", "WD"), threshold = 1,
  max.na.prop = 0.25, use.ff = FALSE)
}
\arguments{
\item{stationObj}{An R-VALUE object containing station data (as returned by \code{\link{loadValueStations}}).}

\item{predictions.file}{Path to the file storing the predictions (passed to \code{\link{loadValuePredictions}}).
Default to NULL, meaning that the correlation matrix is done on the observations.}

\item{season}{Character string indicating the target season. Accepted values are 
\code{c("annual", "DJF", "MAM", "JJA", "SON")}}

\item{aggr.type}{Type of aggregation in the case of multiple realizations. Should the aggregation of 
multiple members be performed \code{"before"} (default) or \code{"after"} computing the correlations?. Ignored
 in the case of observations and deterministic predictions.}

\item{prob.type}{Character vector indicating the type of probability to be computed. Currently accepted values are:
\code{"DD"} for dry-dry, \code{"DW"} for dry-wet, \code{"WW"} for wet-wet and \code{"WD"} for wet-dry.}

\item{threshold}{Threshold above which values are used/discarded (i.e., values greater or equal than \code{threshold} are considered as Wet).}

\item{max.na.prop}{Maximum allowed proportion of missing data (Default to 0.25). See details}

\item{use.ff}{Optional. If set to \code{TRUE}, the loaded stations/predictions array is written to disk using the \pkg{ff} package in order to avoid 
memory problems. Recommended when working with large stochastic prediction datasets in resource-limited machines.}
}
\value{
A 2D matrix. Attributes indicate the station names (in the row/column order they appear), and their
geographical coordinates. In addition, the type of joint probability, as coded in argument \code{prob.type} is indicated in attribute \code{"joint_prob_type"}.
}
\description{
Computes different joint (precipitation) probabilities between stations
}
\author{
J. Bedia
}

