% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadValuePredictions}
\alias{loadValuePredictions}
\title{Load VALUE predictions data}
\usage{
loadValuePredictions(stationObj, predictions.file, tz = "")
}
\arguments{
\item{stationObj}{A station data object returned by \code{\link{loadValueStations}}. See details.}

\item{predictions.file}{path to the file (either text or zip) containing the predictions.}

\item{tz}{Optional. A time zone specification to be used for the conversion of dates. See more details in
\code{\link{loadValueStations}}.}
}
\value{
A predictions object. This is equivalent to the stations object (see \code{\link{loadValueStations}}),
but the data element may vary its shape to include the \code{"member"} dimension in case of stochastic
predictions with several realizations. Also, a global attribute \code{"datatype"} is set, and assigned
the value \code{"predictions"} to differentiate it from observations data.
}
\description{
Load predictions data, either deterministic (one single txt file) or stochastic (several
 realizations stored in sepparate txt files, bundled in a single zip file).
}
\details{
The idea of this function is to rely on the previously loaded observations data, so all
 the information required to define the predictions subset (time and station codes) is retrieved from
 the observations data subset.
}
\seealso{
Other loading: \code{\link{loadValueStations}}
}

