% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.bias.R
\name{measure.bias}
\alias{measure.bias}
\title{Bias}
\usage{
measure.bias(obs, prd)
}
\arguments{
\item{obs}{Mean value for the observed series or a vector with the observed time series.}

\item{prd}{Mean value for the predicted series or a vector with the predicted time series.}
}
\value{
A float number corresponding to the bias.
}
\description{
Function to compute the difference between the observed and predicted means. 
The measure depends on the index mean.
}
\examples{
\dontrun{
# Bias Braganca:
data(precipIberiaECA)
obs <- precipIberiaECA$observations$Data[,1]
prd <- precipIberiaECA$predictions$Data[,1]
# Considering the time series:
measure.bias(obs,prd)
# Considering the index associated with the bias (mean):
obs <- index.mean(obs)
prd <- index.mean(prd)
measure.bias(obs,prd)
}
}
\author{
Daniel San-Mart\'in \email{daniel@predictia.es}, J. Bedia, D. San-Mart\'in, S. Herrera
}

