% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.bias.rel.R
\name{measure.bias.rel}
\alias{measure.bias.rel}
\title{Relative bias}
\usage{
measure.bias.rel(indexObs = NULL, indexPrd = NULL, obs = NULL,
  prd = NULL, dates)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}

\item{dates}{dates}
}
\value{
A float number corresponding to the bias.
}
\description{
Function to compute the difference between the relative bias between the predicted and observed means. 
The measure depends on the index mean.
}
\author{
J. Bedia, D. San-Mart\'in, S. Herrera
}
