% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.bs.R
\name{measure.bs}
\alias{measure.bs}
\title{Brier score}
\usage{
measure.bs(indexObs = NULL, indexPrd = NULL, obs, prd, dates,
  threshold = 0, condition)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}

\item{threshold}{A float number defining the threshold considered. Default to 0 (for temperature).}

\item{condition}{Inequality operator to be applied considering the given threshold.
\code{"GT"} = greater than the value of \code{threshold}, \code{"GE"} = greater or equal,
  \code{"LT"} = lower than, \code{"LE"} = lower or equal than.}
}
\value{
A float number corresponding to the BS.
}
\description{
Function to compute the Brier score.
}
\author{
J. Bedia
}
\references{
https://en.wikipedia.org/wiki/Brier_score
}

