% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.cor.R
\name{measure.cor}
\alias{measure.cor}
\title{Correlation}
\usage{
measure.cor(indexObs = NULL, indexPrd = NULL, obs, prd,
  method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}

\item{method}{Type of correlation applied. Options: \code{c("pearson","kendall","spearman")}.}
}
\value{
A float number corresponding to the correlation coefficient of choice between the predicted and observed series.
}
\description{
Function to compute the temporal correlation between the observed and predicted time series.
}
\author{
J. Bedia, D. San-Martin, S. Herrera
}

