% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.corlongterm.R
\name{measure.corlongterm}
\alias{measure.corlongterm}
\title{Correlation on Long time scales}
\usage{
measure.corlongterm(indexObs = NULL, indexPrd = NULL, obs, prd, dates,
  method = c("pearson", "kendall", "spearman"), t.aggr = "annual",
  timescale = 1, detrend = FALSE, plot = FALSE)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}

\item{dates}{dates}

\item{method}{Character. Type of correlation applied. Options: \code{c("pearson","kendall","spearman")}.}

\item{t.aggr}{Character. Temporal aggregation options. Current accepted values are either \code{"annual"} or \code{"seasonal"}.}

\item{timescale}{Integer. Approximate filtering time scale (according to the \code{t.aggr} units).}

\item{detrend}{Logical. Whether the aggregated time series should be linearly detrended prior to filtering}

\item{plot}{Logical, for internal use only. Should the original (aggregated) and filtered series be plotted?}
}
\value{
A float number corresponding to the correlation coefficient of choice between the predicted and observed series.
}
\description{
Function to compute the temporal correlation between the observed and predicted time series.
}
\author{
D. Maraun, J. Bedia
}
\keyword{internal}
