% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.ks.R
\name{measure.ks}
\alias{measure.ks}
\title{Two-sided KS score}
\usage{
measure.ks(indexObs = NULL, indexPrd = NULL, obs, prd)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}
}
\value{
A float number corresponding to the KS score.
}
\description{
Function to compute the (two-sided) Kolmogorov-Smirnov (KS) score between the observed and predicted time series.
}
\author{
J. Bedia
}
