% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.mae.R
\name{measure.mae}
\alias{measure.mae}
\title{Mean absolute error}
\usage{
measure.mae(obs, prd)
}
\arguments{
\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}
}
\value{
A float number corresponding to the mean absolute difference between the predicted and observed series.
}
\description{
Function to compute the mean absolute difference between the observed and predicted time series.
}
\examples{
\dontrun{
# Mean absolute error of Braganca:
data(precipIberiaECA)
obs <- precipIberiaECA$observations$Data[,1]
prd <- precipIberiaECA$predictions$Data[,1]
measure.mae(obs,prd)
}
}
\author{
Daniel San-Mart\'in \email{daniel@predictia.es}, J. Bedia, D. San-Mart\'in, S. Herrera
}

