% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.ratio.R
\name{measure.ratio}
\alias{measure.ratio}
\title{Ratio}
\usage{
measure.ratio(indexObs = NULL, indexPrd = NULL, obs = NULL, prd = NULL)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}
}
\value{
A float number corresponding to the ratio.
}
\description{
Function to compute the ratio between the predicted and observed means. 
The measure depends on the index mean.
}
\author{
J. Bedia, D. San-Mart\'in, S. Herrera
}

