% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.rmse.R
\name{measure.rmse}
\alias{measure.rmse}
\title{Root mean square error}
\usage{
measure.rmse(obs, prd)
}
\arguments{
\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}
}
\value{
A float number corresponding to the root mean square error between the predicted and observed series.
}
\description{
Function to compute the root mean square error between the observed and predicted time series.
}
\examples{
\dontrun{
# Root mean square error of Braganca:
data(precipIberiaECA)
obs <- precipIberiaECA$observations$Data[,1]
prd <- precipIberiaECA$predictions$Data[,1]
measure.rmse(obs, prd)
}
}
\author{
Daniel San-Mart\'in \email{daniel@predictia.es}, J. Bedia, D. San-Mart\'in, S. Herrera
}

