% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.filter.VALUE.R
\name{na.filter.VALUE}
\alias{na.filter.VALUE}
\title{Remove stations above a given proportion of missing data}
\usage{
na.filter.VALUE(valueObj, max.na.prop = 0)
}
\arguments{
\item{valueObj}{A VALUE R object, but see Details.}

\item{max.na.prop}{Maximum allowed proportion of missing data (Default to 0, i.e., no missing values are allowed)}
}
\value{
A R.VALUE object without the stations above the user-defined threshold of missing data
}
\description{
Remove stations over a user-defined proportion of missing data from a R.VALUE multi-station object,
}
\details{
The function assumes that all dimensions are present (i.e., function dimFix has been applied)
}
\author{
J Bedia
}
