% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFUN.R
\name{subsetData}
\alias{subsetData}
\title{subsetData}
\usage{
subsetData(obj, startDates, endDates, stations)
}
\arguments{
\item{startDates}{Start dates}

\item{endDates}{End dates}

\item{stations}{Stations}

\item{data}{Value object of observations/predictions}
}
\value{
A list with obj and pred intersected
}
\description{
Time and stations subsetting
}
\author{
S. Herrera
}
\keyword{internal}
