% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetVALUE.R
\name{subsetVALUE}
\alias{subsetVALUE}
\title{Select an arbitrary subset from a VALUE object along one or more of its dimensions}
\usage{
subsetVALUE(valueObj, stationID = NULL, members = NULL, season = NULL,
  dates = NULL, years = NULL, lonLim = NULL, latLim = NULL)
}
\arguments{
\item{valueObj}{The input object to be subset. This is either an observation, as returned by \code{loadValueStations}, 
, or a predictions object as returned by \code{loadValuePredictions}.}

\item{stationID}{Character string. Station codes.}

\item{members}{An integer vector indicating \strong{the position} of the members to be subset. Default to NULL.}

\item{season}{An integer vector indicating the months to be subset.}

\item{dates}{A vector of dates for subsetting on arbitrary dates. This is a character (or \code{POSIXct}) vector
following the format \dQuote{YYYY-MM-DD} (i.e., \code{format = "\%Y-\%m-\%d"} as in \code{\link{strptime}}). 
This is directly passed by the element \code{valueObject$Dates$start}, for instance.}

\item{years}{The years to be selected. Note that this can be either a continuous or discontinuous
series of years, the latter option often used in a cross-validation framework.
 See details for year-crossing seasons. Default to \code{NULL} (no subsetting is performed on the time dimension).}

\item{lonLim}{Vector of length = 2, with minimum and maximum longitude coordinates, in decimal degrees,
of the bounding box defining the subset. For single-station subsets, a numeric value with the
longitude coordinate. If \code{NULL} (default), no subsetting is performed on the longitude dimension}

\item{latLim}{Same as \code{lonLim} argument, but for latitude.}
}
\value{
A new VALUE stations/predictions object that is a logical subset of the input object along the specified dimensions.
}
\description{
Creates a new predictions/observations VALUE object that is a subset of the input object 
along the selected dimensions
}
\details{
The attribute \code{subset} will be added to the different slots corresponding to the subset dimensions, taking
the value of the subroutine called in each case (e.g.: attribute subset will have the value \code{subsetSpatial}
in the xyCoords slot after spatial subsetting...).

\strong{Time slicing by years}

In case of year-crossing seasons (e.g. boreal winter (DJF), \code{season = c(12,1,2)}),
the season is assigned to the years of January and February 
(i.e., winter of year 2000 corresponds to Dec 1999, Jan 2000 and Feb 2000). Thus, 
the \code{years} argument must be introduced accordingly.

 \strong{Spatial slicing}
 
 Spatial subset definition is done via the \code{lonLim} and \code{latLim} arguments, in the same way as
  in the \code{\link{loadValueStations}} function. It is also possible to
  make single-station selections from a multi-station object, just by specifying a single coordinate instead of a range
   as the argument value. For instance \code{lonLim=c(-10,10)} and \code{latLim=c(35,45)} indicates a
 rectangular window centered in the Iberian Peninsula), and single grid-cell values
 (for instance \code{lonLim=-3.21} and \code{latLim=41.087} for retrieving the data in the closest grid
 point to the point coordinate -3.21E, 41.087N. In both cases, the function
 operates by finding the nearest (euclidean distance) stations to the coordinates introduced
  (either north/south or east/westward).
}
\author{
J. Bedia
}

