% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tasmaxIberiaECA}
\alias{tasmaxIberiaECA}
\title{Observations and predictions for max surface temperature in Iberia}
\format{A list containing two objects of the types (not formally classes) \code{stations.observations}
 and \code{stations.predictions}}
\description{
The object contains two objects, \code{observations} and \code{predictions}
of daily maximum near-surface air temperature in summer (JJA) for the period 2001-2010,
in a set of 9 ECA stations from the VALUE reference dataset in the Iberian Peninsula.
}
\details{
The predictions have been calculated using NCEP fields of mean near-surface temperature and sea-level
pressure after PCA analysis as predictors, and the analog method considering the closest neighbor in euclidean
 space, for the training period 1981-2000. Observations data were obtained from the \pkg{R_VALUE} built-in
 dataset \code{VALUE_ECA.zip}. NCEP data and analog method implementation is available in the \pkg{downscaleR}
 package.
}
\examples{
data(tasmaxIberiaECA)
plot(tasmaxIberiaECA$predictions$Data[ ,2], tasmaxIberiaECA$observations$Data[ ,2],
   asp = 1, ylab = "Obs", xlab = "Pred")
title(tasmaxIberiaECA$observations$Metadata$longname[2])
mtext("Tmax JJA (2001-2010)")
r2 <- format(summary(lm(tasmaxIberiaECA$predictions$Data[ ,2] ~
   tasmaxIberiaECA$observations$Data[ ,2]))$adj.r.squared, digits = 2)
text(20,40, paste("Rsq =", r2))
abline(lm(tasmaxIberiaECA$predictions$Data[ ,2] ~ tasmaxIberiaECA$observations$Data[ ,2]))
}
\references{
\url{https://github.com/SantanderMetGroup/downscaleR/wiki/analog_demo}
}

