% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validation}
\alias{validation}
\title{Validation}
\usage{
validation(obs, prd, lag.max = 3, lowVarPeriod = 1, Nbins = 100)
}
\arguments{
\item{obs}{}

\item{prd}{}
}
\value{
A validation object.
}
\description{
Calculation of the indicators and scores defined for each variable in the VALUE Action Cost.
}
\examples{
#Example predictions
obs.dataset <- file.path(find.package("R.VALUE"), "example-observations.zip")
obs <- loadValueStations(obs.dataset, "tmin", season = 6:8, years = 2001)
# Loading deterministic predictions
pred.file1 <- file.path(find.package("R.VALUE"), "example-prediction.txt")
pred <- loadValuePredictions(obs, pred.file1)
str(pred$Data) # 2D array
# Loading stochastic predictions (several realizations)
pred.file2 <- file.path(find.package("R.VALUE"), "example-prediction-multimember.zip")
pred2 <- loadValuePredictions(obs, pred.file2)
str(pred2$Data) # 3D array with 'member' dimension
}

