% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{get98th}
\alias{get98th}
\title{98th percentile}
\usage{
get98th(data, MARGIN)
}
\description{
Get annual and seasonal 98th percentile from a station or field object
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg} and J. Bedia
}
\keyword{internal}

