% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{getACF}
\alias{getACF}
\title{Autocorrelation}
\usage{
getACF(data, lag.max = 3)
}
\description{
Get annual and seasonal Autocorrelation at pre-defined lags from a station or field object
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, S. Herrera and J. Bedia
}
\keyword{internal}

