% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{getCM}
\alias{getCM}
\title{Get annual and seasonal Cramer von Misses index}
\usage{
getCM(dataRef, data, Nbins = 100)
}
\author{
Ole Roessler \email{ole.roessler@giub.unibe.ch} and J. Bedia and S. Herrera
}
\keyword{internal}

