% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFreqDD}
\alias{getFreqDD}
\title{Dry-dry probability}
\usage{
getFreqDD(data, threshold)
}
\description{
Get annual and seasonal dry-dry-probability from a station or field object
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg} and J. Bedia and S. Herrera
}
\keyword{internal}

