% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{getFreqDW}
\alias{getFreqDW}
\title{Dry-wet probability}
\usage{
getFreqDW(data, threshold)
}
\description{
Get annual and seasonal dry-wet-probability from a station or field object
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg} and J. Bedia and S. Herrera
}
\keyword{internal}

