% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{getFreqGT}
\alias{getFreqGT}
\title{Get annual and seasonal amount of days above of a predefined threshold from a station or field object}
\usage{
getFreqGT(data, threshold, MARGIN = NULL)
}
\author{
Dougals Maraun \email{dmaraun@geomar.de} and J. Bedia
}
\keyword{internal}

