% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{getFreqLT}
\alias{getFreqLT}
\title{Get annual and seasonal amount of days below of a predefined threshold from a station or field object}
\usage{
getFreqLT(data, threshold, MARGIN = NULL)
}
\author{
Dougals Maraun \email{dmaraun@geomar.de} and J. Bedia
}
\keyword{internal}

