% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFreqLT}
\alias{getFreqLT}
\title{Get annual and seasonal amount of days below of a predefined threshold from a station or field object}
\usage{
getFreqLT(data, threshold, MARGIN = NULL)
}
\description{
Get annual and seasonal amount of days below of a predefined threshold from a station or field object
}
\author{
Dougals Maraun \email{dmaraun@geomar.de} and J. Bedia
}
\keyword{internal}

