% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFreqWD}
\alias{getFreqWD}
\title{Wet-dry probability}
\usage{
getFreqWD(data, threshold)
}
\description{
Get annual and seasonal wet-dry-probability from a station or field object
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg} and J. Bedia
}
\keyword{internal}

