% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{getFreqWW}
\alias{getFreqWW}
\title{Wet-wet probability}
\usage{
getFreqWW(data, threshold)
}
\description{
Get annual and seasonal wet-wet-probability from a station or field object
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg} and J. Bedia and S. Herrera
}
\keyword{internal}

