% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFUN.R
\name{getIntersect}
\alias{getIntersect}
\title{getIntersect}
\usage{
getIntersect(obs, prd)
}
\arguments{
\item{obs}{Value object of observations}

\item{prd}{Value object of predictions}
}
\value{
A list with obj and pred intersected
}
\description{
Temporal and spatial matching between obs and pred
}
\details{
The function ensures that all records are in choronological order, by reordering -if necessary- the time dimension.
Dates are ordered as characters, without performing the conversion to date object, which has proven time-consuming when 
repeatedly evaluated.
}
\author{
S. Herrera, D. San-Martin, J Bedia
}
\keyword{internal}

