% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getSeason.VALUE}
\alias{getSeason.VALUE}
\title{Get season from a station or field object}
\usage{
getSeason.VALUE(objDates)
}
\arguments{
\item{objDates}{Dates as directly extracted from any station or predictions object}
}
\value{
An integer vector with the season
}
\description{
Retrieves the season encompassed by a station or field object
}
\author{
juaco
}
\keyword{internal}
