% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{getSkew}
\alias{getSkew}
\title{Get annual and seasonal skewness from a station or field object}
\usage{
getSkew(data, MARGIN)
}
\author{
Ole Roessler \email{ole.roessler@giub.unibe.ch} and J. Bedia
}
\keyword{internal}

