% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getVar}
\alias{getVar}
\title{Get annual and seasonal variance from a station or field object}
\usage{
getVar(data, MARGIN)
}
\description{
Get annual and seasonal variance from a station or field object
}
\author{
Ole Roessler \email{ole.roessler@giub.unibe.ch} and J. Bedia
}
\keyword{internal}

