% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validationUtil.R
\name{getVarLF}
\alias{getVarLF}
\title{Get annual and seasonal proportion of variance in low frequency}
\usage{
getVarLF(data, lowVarPeriod, INDEX = 1:dim(data)[1])
}
\description{
Proportion of variance in low frequency
}
\author{
Dougals Maraun \email{dmaraun@geomar.de} and J. Bedia and S. Herrera
}
\keyword{internal}

