% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWDsld}
\alias{getWDsld}
\title{Get annual and seasonal above of a predefined threshold Spell Length Distribution from a station or field object}
\usage{
getWDsld(data, threshold, INDEX = 1:dim(data)[1])
}
\description{
Get annual and seasonal above of a predefined threshold Spell Length Distribution from a station or field object
}
\keyword{internal}

