% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWet98th}
\alias{getWet98th}
\title{98th percentile}
\usage{
getWet98th(data, threshold, MARGIN = 1:length(dim(data)))
}
\description{
Get annual and seasonal 98th percentile from a station or field object
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg} and J. Bedia
}
\keyword{internal}

