% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.acf.R
\name{index.acf}
\alias{index.acf}
\title{Autocorrelation function}
\usage{
index.acf(ts, lag.max = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{lag.max}{Maximum lag considered for acf calculation. Default \code{lag.max = 1}.}
}
\value{
A scalar with the estimated autocorrelation for that lag.
}
\description{
Function to compute the estimated autocorrelation function at pre-defined lags.
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, J. Bedia, D. San-Mart\'in, S. Herrera
}
