% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.amountFreqGT.R
\name{index.amountFreqGT}
\alias{index.amountFreqGT}
\title{Amount from events above threshold}
\usage{
index.amountFreqGT(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the input time series}

\item{threshold}{A float number defining the threshold considered. Default to 1.}
}
\value{
A float number corresponding to the amount of days above the defined threshold of the input.
}
\description{
Function to compute the amount from events above a predefined threshold.
}
\examples{
\dontrun{
# Precipitation amounts in Braganca:
data(precipIberiaECA)
obs <- precipIberiaECA$observations$Data[,1]
# Above default threshold (1 mm)
index.amountFreqGT(obs)
# Above threshold of 10 mm
index.amountFreqGT(obs, 10)
}
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, J. Bedia, D. San-Mart\'in, S. Herrera
}

