% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.annual.cycle.filtered.R
\name{index.annual.cycle.filtered}
\alias{index.annual.cycle.filtered}
\title{Filtered annual cycle maxima statistics}
\usage{
index.annual.cycle.filtered(ts, dates, type = "phase", peak = 1,
  thresh = 0.1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{dates}{A character (or \code{POSIXct}) vector following the format \dQuote{YYYY-MM-DD}
(i.e., \code{format = "\%Y-\%m-\%d"} as in \code{\link{strptime}}). This is directly passed by the
 VALUE objects (either stations or predictions) through the element \code{object$Dates$start}.}

\item{type}{A character string indicating max \code{"phase"} (days, default), \code{"amp"} (amplitude) or \code{"amprel"} (relative amplitude)}

\item{peak}{= 1 or 2 (maximum or secondary maximum - higher order peaks are ignored)}

\item{thresh}{Threshold that defines weak and double maxima (relative to amplitude of annual cycle)}
}
\value{
A float number corresponding to the amplitude or the (phase) day of the maxima
}
\description{
Returns the amplitude or the location of the maxima of the annual cycle
}
\author{
D. Maraun, J. Bedia, D. San-Martin, J.M. Gutierrez
}
\keyword{internal}
