% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freqGE.R
\name{index.freqGE}
\alias{index.freqGE}
\title{Frequency above or equal the threshold}
\usage{
index.freqGE(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered. Default to 1.}
}
\value{
A float number corresponding to the amount of days above or equal the defined threshold of the input.
}
\description{
Function to compute the amount of days above or equal of a predefined threshold mean index. Missing values are ignored.
}
\author{
Dougals Maraun \email{dmaraun@geomar.de}, J. Bedia, D. San-Mart\'in, S. Herrera
}

