% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freqLE.R
\name{index.freqLE}
\alias{index.freqLE}
\title{Absolute frequency lower or equal the threshold}
\usage{
index.freqLE(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered. Default to 1}
}
\value{
A float number corresponding to the amount of days below or equal the defined threshold of the input.
}
\description{
Function to compute the amount of days below or equal of a predefined threshold mean index.
}
\author{
Douglas Maraun \email{dmaraun@geomar.de}
}

