% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.freqWW.R
\name{index.freqWW}
\alias{index.freqWW}
\title{Wet-wet probability}
\usage{
index.freqWW(ts, threshold = 1)
}
\arguments{
\item{ts}{A vector containing the data}

\item{threshold}{A float number defining the threshold considered. Default to 1.}
}
\value{
A float number corresponding to the Wet-wet transition probability.
}
\description{
Function to compute the wet-wet-probability index.
}
\author{
Neyko Neykov \email{neyko.neykov@meteo.bg}, J. Bedia, D. San-Mart\'in, S. Herrera
}
