% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.rv.ub.R
\name{index.rv.ub}
\alias{index.rv.ub}
\title{Return Value upper tail}
\usage{
index.rv.ub(ts, prob = 20, annual.index = TRUE)
}
\arguments{
\item{ts}{A vector containing the data}

\item{prob}{Return period considered. Default to 20.}

\item{annual.index}{Logical. If set to TRUE, a vector defining the years for aggregation is passed by \code{\link{wrapperFUN}}.
Otherwise no temporal aggregation of the input time series is performed.}
}
\value{
Return value for the upper tail
}
\description{
Function to compute the 1/prob return value (right tail).
}
\author{
Neyko Neykov, S. Herrera, J. Bedia
}

