% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.bias.R
\name{measure.bias}
\alias{measure.bias}
\title{Bias}
\usage{
measure.bias(indexObs = NULL, indexPrd = NULL, obs = NULL, prd = NULL,
  dates)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}
}
\value{
A float number corresponding to the bias.
}
\description{
Function to compute the difference between the observed and predicted means. 
The measure depends on the index mean.
}
\author{
J. Bedia, D. San-Mart\'in, S. Herrera
}

