% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.cm.R
\name{measure.cm}
\alias{measure.cm}
\title{Cramer von Misses}
\source{
L. Sachs and J. Hedderich (2006). Angewandte Statistik. Springer.
}
\usage{
measure.cm(obs, prd, Nbins = 100)
}
\arguments{
\item{obs}{Vector of observations}

\item{prd}{Vector os predictions}

\item{Nbins}{Number of bins used for discretization. Default to 100.}
}
\value{
List with the Cramer von Misses index (\code{CvM}), the p-value (\code{pvalue}) of the test 
and the maximum absolute difference between both the observed and predicted distributions (\code{KSdiff}).
}
\description{
Function to compare the observed and predicted distribution using the Cramer von Misses index.
}
\examples{
\dontrun{
# Cramer von Misses test for Braganca:
data(precipIberiaECA)
obs <- precipIberiaECA$observations$Data[,1]
prd <- precipIberiaECA$predictions$Data[,1]
measure.cm(obs,prd)
}
}
\author{
Ole Roessler \email{ole.roessler@giub.unibe.ch}, J. Bedia, D. San-Mart\'in, S. Herrera
}

