% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.cm.R
\name{measure.cm}
\alias{measure.cm}
\title{Cramer von Misses}
\source{
L. Sachs and J. Hedderich (2006). Angewandte Statistik. Springer.
}
\usage{
measure.cm(indexObs = NULL, indexPrd = NULL, obs = NULL, prd = NULL,
  Nbins = 100)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}

\item{Nbins}{Number of bins used for discretization. Default to 100.}
}
\value{
List with the Cramer von Misses index (\code{CvM}), the p-value (\code{pvalue}) of the test 
and the maximum absolute difference between both the observed and predicted distributions (\code{KSdiff}).
}
\description{
Function to compare the observed and predicted distribution using the Cramer von Misses index.
}
\author{
Ole Roessler \email{ole.roessler@giub.unibe.ch}, J. Bedia, D. San-Mart\'in, S. Herrera
}

