% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.corr.R
\name{measure.cor}
\alias{measure.cor}
\title{Correlation}
\usage{
measure.cor(indexObs = NULL, indexPrd = NULL, obs = NULL, prd = NULL,
  method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{Mean value for the observed series or a vector with the observed time series.}

\item{prd}{Mean value for the predicted series or a vector with the predicted time series.}

\item{method}{Type of correlation applied. Options: "pearson", "kendall" and "spearman".}
}
\value{
A float number corresponding to the correlation coefficient of choice between the predicted and observed series.
}
\description{
Function to compute the temporal correlation between the observed and predicted time series.
}
\author{
Daniel San-Mart\'in \email{daniel@predictia.es}, J. Bedia, D. San-Mart\'in, S. Herrera
}

