% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.mae.R
\name{measure.mae}
\alias{measure.mae}
\title{Mean absolute error}
\usage{
measure.mae(indexObs = NULL, indexPrd = NULL, obs = NULL, prd = NULL,
  dates)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}

\item{dates}{dates}
}
\value{
A float number corresponding to the mean absolute difference between the predicted and observed series.
}
\description{
Function to compute the mean absolute difference between the observed and predicted time series.
}
\author{
Daniel San-Mart\'in \email{daniel@predictia.es}, J. Bedia, D. San-Mart\'in, S. Herrera
}
