% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.rmse.R
\name{measure.rmse}
\alias{measure.rmse}
\title{Root mean square error}
\usage{
measure.rmse(indexObs = NULL, indexPrd = NULL, obs = NULL, prd = NULL,
  dates)
}
\arguments{
\item{indexObs}{index computed from the observations}

\item{indexPrd}{index computed from the predictions}

\item{obs}{A vector of observations}

\item{prd}{A vector of predictions}
}
\value{
A float number corresponding to the root mean square error between the predicted and observed series.
}
\description{
Function to compute the root mean square error between the observed and predicted time series.
}
\author{
Daniel San-Mart\'in \email{daniel@predictia.es}, J. Bedia, D. San-Mart\'in, S. Herrera
}

