% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFUN.R
\name{preprocessVALUE}
\alias{preprocessVALUE}
\title{preprocessVALUE}
\usage{
preprocessVALUE(obs, pred, dates.obs, dates.pred, na.prop)
}
\arguments{
\item{obs}{A time series (vector) of observations}

\item{pred}{A time series (vector) of predictions}

\item{dates.obs}{Calendar dates corresponding to the values in \code{obs}}

\item{dates.pred}{Calendar dates corresponding to the values in \code{pred}}

\item{na.prop}{Maximum allowable proportion of missing values.}
}
\value{
A list with preprocessed \code{obs}, \code{pred} and \code{dates}, as passed to the index.* routines
}
\description{
Preprocessing of input data for index calculation routines
}
\details{
The function performs missing data filtering, temporal matching and (optionally) seasonal subsetting
}
\author{
J. Bedia
}
\keyword{internal}

