% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validation}
\alias{validation}
\title{Validation}
\usage{
validation(obs, prd, lag.max = 3, lowVarPeriod = 1, Nbins = 100,
  prob = 1/20)
}
\arguments{
\item{obs}{A field or station data containing the observed climate data}

\item{prd}{A field containing the simulated climate by the model}

\item{lag.max}{Optional. Maximum lags considered for the autocorrelation. Default = 3}

\item{lowVarPeriod}{Optional. Low pass filter. Default = 1}

\item{Nbins}{Optional. Number of intervals considered in the calculation of the Cramer Von Misses. Default = 100}

\item{prob}{Optional. Inverse of the return period. Default = 1/20 corresponding with the 20 years return value.}
}
\value{
A validation object. A 4-D array (season, station, member, index) with the indices and scores obtained for each season (DJF, MAM, JJA, SON and Annual), location and member.
}
\description{
Function to compute VALUE indices and scores for each variable.
}

