% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFUN.R
\name{wrapperFUN}
\alias{wrapperFUN}
\title{Wrapper function to launch the validation}
\usage{
wrapperFUN(metric = c("obs", "pred", "measure"), names = NULL,
  season = c("annual", "DJF", "MAM", "JJA", "SON"), aggr.type = c("before",
  "after"), aggr.fun = "mean", index.fun = NULL, measure.fun = NULL,
  index.args = NULL, measure.args = NULL, o = o, p = p, na.prop = 0.9)
}
\arguments{
\item{metric}{Character vector.}

\item{names}{Character vector of the same length than \code{metric}. Names of the indices/measures to be applied}

\item{season}{Character vector defining the target season(s). Default to annual + 4 standard seasons.}

\item{aggr.type}{Character vector of length one. Should member aggregation be performed \code{"before"} or \code{"after"} the validation?.
Ignored for observations and deterministic predictions.}

\item{aggr.fun}{Character vector of length one. Multimember aggregation function. Default to \code{"mean"}.}

\item{index.fun}{A character vector with the name of the R function that computes the index.}

\item{measure.fun}{A character vector with the name of the R function that computes the measure.}

\item{index.args}{A list with additional arguments passed to \code{index.fun}. It contains a key-value list for each additional argument.}

\item{measure.args}{Same as \code{index.args} but for the measure function.}

\item{o}{R object containing the observations as returned by \code{\link{loadValueStations}}.}

\item{p}{R object containing the predictions as returned by \code{\link{loadValuePredictions}}.}

\item{na.prop}{Maximum allowable proportion of missing data. Default to 0.9}
}
\value{
A 3D array with labelled dimensions station, season and metric
}
\description{
Launches the VALUE validation framework according to the arguments passed by the
}
\examples{
\dontrun{
# Load observations
obs.file <- file.path(find.package("R.VALUE"), "example_datasets", "VALUE_ECA_86_v2.zip")
o <- loadValueStations(obs.file, var = "tmin")
prdfile <- list.files(file.path(find.package("R.VALUE"), "example_datasets"),
                      pattern = "predictions_portal_exp1a_deterministic",
                      full.names = TRUE)
# Load predictions
p <- loadValuePredictions(o, predictions.file = prdfile)
# Create argument list for computing mean bias:
c1 = list()
c1$metric = c("obs", "pred", "measure")
c1$names = c("obsMean", "predMean","meanBias")
c1$season = c("annual", "DJF", "MAM", "JJA", "SON")
c1$aggr.type = "after"
c1$aggr.fun = "mean"
c1$index.fun = "index.mean.R"
c1$measure.fun = "measure.bias.R"
c1$index.args = NULL
c1$measure.args = NULL
c1$o = o
c1$p = p
c1$na.prop = .9

# Call validation wrapper
a <- do.call("wrapperFUN", c1)
str(a)
}
}
\author{
J. Bedia, D. San Martin
}

