% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFUN.R
\name{wrapperFUN}
\alias{wrapperFUN}
\title{Wrapper function to launch the validation}
\usage{
wrapperFUN(metric = c("obs", "pred", "measure"), names = NULL,
  season = c("annual", "DJF", "MAM", "JJA", "SON"),
  member.aggregation = "none", n.mem = NULL, index.fun = NULL,
  measure.fun = NULL, index.args = NULL, measure.args = NULL, o = o,
  p = p, processes = data.frame(), processNames = c(), na.prop = 1)
}
\arguments{
\item{metric}{Character vector.}

\item{names}{Character vector of the same length than \code{metric}. Names of the indices/measures to be applied}

\item{season}{Character vector defining the target season(s). Default to annual + 4 standard seasons.}

\item{member.aggregation}{Character vector of length one. What aggregation function should be applied to multipe realizations
before computing the indices?. Default to \code{"none"}, meaning that the indices are computed in a member-wise basis, and only
 after that the index values are aggregated to compute the measure. The only additional option currently used
  is \code{"mean"}, for cases when the realizations are averaged before computing the index. 
 Ignored for observations and deterministic predictions.}

\item{n.mem}{Number of members to be considered for validation. Default (\code{n.mem = NULL}) to all members,
(or 1 in case of deterministic predictions). Note that the number of members selected are used for \code{member.aggregation},
independently of whether this aggregation is performed before or after.}

\item{index.fun}{A character vector with the name of the R function that computes the index.}

\item{measure.fun}{A character vector with the name of the R function that computes the measure.}

\item{index.args}{A list with additional arguments passed to \code{index.fun}. It contains a key-value list for each additional argument.}

\item{measure.args}{Same as \code{index.args} but for the measure function.}

\item{o}{R object containing the observations as returned by \code{\link{loadValueStations}}.}

\item{p}{R object containing the predictions as returned by \code{\link{loadValuePredictions}}.}

\item{processes}{processes}

\item{processNames}{Labels identifying the processes}

\item{na.prop}{Maximum allowable proportion of missing data. Default to 0.9}
}
\value{
A 3D array with labelled dimensions station, season and metric
}
\description{
Launches the VALUE validation framework according to the arguments passed by the database.
}
\note{
This function is not envisaged to be directly called by the user. It is internally called by the validation portal.
}
\examples{
\dontrun{
# Load observations
obs.file <- file.path(find.package("R.VALUE"), "example_datasets", "VALUE_ECA_86_v2.zip")
o <- loadValueStations(obs.file, var = "tmin")
prdfile <- list.files(file.path(find.package("R.VALUE"), "example_datasets"),
                      pattern = "example_predictions_tmin_portal_exp1a_deterministic",
                      full.names = TRUE)
# Load predictions
p <- loadValuePredictions(o, predictions.file = prdfile)
# Example: computing correlation. Seasonal cycle removal is applied
vno <- wrapperFUN(metric = c('measure'),
                  names = c('measure'),
                  season = c('annual','DJF','MAM','JJA','SON'),
                  member.aggregation = 'none',
                  index.fun = NULL,
                  measure.fun = 'measure.cor.R',
                  index.args = NULL,
                  measure.args = list('method' = 'pearson','deseason' = TRUE),
                  o = o,
                  p = p)
str(vno)
}
}
\author{
J. Bedia, D. San Martin, M. Tuni
}

